+++ app/assets/javascripts/discourse-common/addon/helpers/d-icon.js
@@ -1,9 +1,9 @@
 import { htmlSafe } from "@ember/template";
-import { registerUnbound } from "discourse-common/lib/helpers";
+import { registerRawHelper } from "discourse-common/lib/helpers";
 import { renderIcon } from "discourse-common/lib/icon-library";
 
 export default function icon(id, options = {}) {
   return htmlSafe(renderIcon("string", id, options));
 }
 
-registerUnbound("d-icon", icon);
+registerRawHelper("d-icon", icon);
