+++ app/assets/javascripts/discourse/app/helpers/format-username.js
@@ -1,4 +1,5 @@
 import { formatUsername } from "discourse/lib/utilities";
-import { registerUnbound } from "discourse-common/lib/helpers";
+import { registerRawHelper } from "discourse-common/lib/helpers";
 
-export default registerUnbound("format-username", formatUsername);
+export default formatUsername;
+registerRawHelper("format-username", formatUsername);
