+++ app/assets/javascripts/discourse/app/helpers/theme-prefix.js
@@ -1,0 +1,6 @@
+import { registerRawHelper } from "discourse-common/lib/helpers";
+
+registerRawHelper("theme-prefix", themePrefix);
+export default function themePrefix(themeId, key) {
+  return `theme_translations.${themeId}.${key}`;
+}
