+++ app/assets/javascripts/discourse/app/helpers/topic-featured-link.js
@@ -1,7 +1,8 @@
 import { htmlSafe } from "@ember/template";
 import renderTopicFeaturedLink from "discourse/lib/render-topic-featured-link";
-import { registerUnbound } from "discourse-common/lib/helpers";
+import { registerRawHelper } from "discourse-common/lib/helpers";
 
-export default registerUnbound("topic-featured-link", function (topic, params) {
+registerRawHelper("topic-featured-link", topicFeaturedLink);
+export default function topicFeaturedLink(topic, params) {
   return htmlSafe(renderTopicFeaturedLink(topic, params));
-});
+}
